ci(bench): add areaY knob + make PDR accumulators explicit doubles (#24)
Expose an areaY workflow input (blank = preset 300) so a guaranteed
in-range single-hop sanity anchor can be run (small areaX+areaY): the
calc must read ~100% there, proving the PDR formula empirically rather
than by inspection.

Also make the FlowMonitor tx/rx accumulators doubles in manet-baselines.
The ratio already did float division (100.0 * promotes before the
divide), so this is cosmetic and does not change any result — it just
makes the accumulator types explicit.

+++ ns3/examples/manet-baselines.cc
@@ -159,7 +159,9 @@ Result RunOne(const std::string& proto, const Params& P, uint32_t seed) {
 
     Result r;
     r.proto = proto;
-    uint64_t tx = 0, rx = 0;
+    // tx/rx as doubles: the ratio below already does float division (100.0 *
+    // promotes first), this just keeps the accumulators explicit.
+    double tx = 0, rx = 0;
     double totalDelay = 0.0;
     std::map<uint32_t, uint64_t> delayBins;
     double binWidth = 0.0;
